Fixing per codeclimate feedback
